♻️ Refactor `get_chat`
