+++ plugins/system/update.py
@@ -7,7 +7,7 @@ from telegram import Message, Update
 from telegram.error import BadRequest, Forbidden
 from telegram.ext import CallbackContext
 
-from core.builtins.contexts import BotContext
+from core.builtins.contexts import ApplicationContext
 from core.plugin import Plugin, handler
 from utils.decorators.admins import bot_admins_rights_check
 from utils.helpers import execute
@@ -33,7 +33,7 @@ class UpdatePlugin(Plugin):
             async with async_open(UPDATE_DATA) as file:
                 data = jsonlib.loads(await file.read())
             try:
-                reply_text = Message.de_json(data, BotContext.get().tg_app.bot)
+                reply_text = Message.de_json(data, ApplicationContext.get().tg_app.bot)
                 await reply_text.edit_text("重启成功")
             except (BadRequest, Forbidden, KeyError) as exc:
                 logger.error("UpdatePlugin 编辑消息出现错误")
